+++ src/fuzz_introspector/analyses/optimal_targets.py
@@ -24,14 +24,83 @@ from typing import (
 
 from fuzz_introspector import analysis
 from fuzz_introspector import constants
-from fuzz_introspector import data_loader
 from fuzz_introspector import html_report
 from fuzz_introspector import html_helpers
 from fuzz_introspector import utils
 from fuzz_introspector.datatypes import (project_profile, fuzzer_profile,
                                          function_profile)
+from fuzz_introspector.exceptions import DataLoaderError
 
 logger = logging.getLogger(name=__name__)
+
+
+def add_func_to_reached_and_clone(
+    merged_profile_old: project_profile.MergedProjectProfile,
+    func_to_add: function_profile.FunctionProfile
+) -> project_profile.MergedProjectProfile:
+    """
+    Add new functions as "reached" in a merged profile, and returns
+    a new copy of the merged profile with reachability information as if the
+    functions in func_to_add are added to the merged profile.
+    The use of this is to calculate what the state will be of a merged profile
+    by targetting a new set of functions.
+    We can use this function in a computation of "optimum fuzzer target analysis", which
+    computes what the combination of ideal function targets.
+    """
+    logger.info("Creating a deepcopy")
+    merged_profile = copy.deepcopy(merged_profile_old)
+
+    # Update hitcount of the function in the new merged profile
+    logger.info("Updating hitcount")
+    f = merged_profile.all_functions[func_to_add.function_name]
+    if f.cyclomatic_complexity == func_to_add.cyclomatic_complexity:
+        f.hitcount = 1
+
+    # Update hitcount of all functions reached by the function
+    for func_name in func_to_add.functions_reached:
+        if func_name not in merged_profile.all_functions:
+            if merged_profile_old.profiles[0].target_lang == "jvm":
+                logger.debug(f"{func_name} not provided within classpath")
+            else:
+                logger.error(f"Mismatched function name: {func_name}")
+            continue
+        f = merged_profile.all_functions[func_name]
+        f.hitcount += 1
+
+        f.reached_by_fuzzers.append(
+            utils.demangle_cpp_func(func_to_add.function_name))
+
+    # Recompute all analysis that is based on hitcounts in all functions as hitcount has
+    # changed for elements in the dictionary.
+    logger.info("Updating hitcount-related data")
+    for f_profile in merged_profile.all_functions.values():
+        cc = 0
+        uncovered_cc = 0
+        for reached_func_name in f_profile.functions_reached:
+            if reached_func_name not in merged_profile.all_functions:
+                if merged_profile_old.profiles[0].target_lang == "jvm":
+                    logger.debug(
+                        f"{reached_func_name} not provided within classpath")
+                else:
+                    logger.error(
+                        f"Mismatched function name: {reached_func_name}")
+                continue
+            f_reached = merged_profile.all_functions[reached_func_name]
+            cc += f_reached.cyclomatic_complexity
+            if f_reached.hitcount == 0:
+                uncovered_cc += f_reached.cyclomatic_complexity
+
+        # set complexity fields in the function
+        f_profile.new_unreached_complexity = uncovered_cc
+        if f_profile.hitcount == 0:
+            f_profile.new_unreached_complexity += f_profile.cyclomatic_complexity
+        f_profile.total_cyclomatic_complexity = cc + f_profile.cyclomatic_complexity
+
+    if merged_profile.all_functions[func_to_add.function_name].hitcount == 0:
+        logger.info("Error. Hitcount did not get set for some reason. Exiting")
+        raise DataLoaderError("Hitcount did not get set for some reason")
+
+    return merged_profile
 
 
 class OptimalTargets(analysis.AnalysisInterface):
@@ -198,7 +267,7 @@ class OptimalTargets(analysis.AnalysisInterface):
             optimal_func = sorted_by_undiscovered_complexity[0]
             optimal_functions_targeted.append(optimal_func)
 
-            new_merged_profile = data_loader.add_func_to_reached_and_clone(
+            new_merged_profile = add_func_to_reached_and_clone(
                 new_merged_profile, optimal_func)
 
             # Update the optimal targets. We only need to do this
